Revert "kenai completed"
