+++ src/config/settings.py
@@ -53,7 +53,6 @@ INSTALLED_APPS = [
     "sos",
     "forum",
     "common",
-    "kenai",
 ]
 
 MIDDLEWARE = [
@@ -254,5 +253,3 @@ if os.environ.get("SECURE_SSL_REDIRECT"):
     # SECURE_SSL_REDIRECT = os.environ.get("SECURE_SSL_REDIRECT")
     # SESSION_COOKIE_SECURE = os.environ.get("SESSION_COOKIE_SECURE")
     # CSRF_COOKIE_SECURE = os.environ.get("CSRF_COOKIE_SECURE")
-
-OPENAI_API_KEY = os.environ.get("OPENAI_API_KEY")
