+++ src/config/urls.py
@@ -43,5 +43,4 @@ urlpatterns = [
     path("v1/sos/", include("sos.urls")),
     path("v1/forum/", include("forum.urls")),
     path("v1/common/", include("common.urls")),
-    path("v1/kenai/", include("kenai.urls")),
 ]
