--- src/kenai/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import Conversation
-
-admin.site.register(Conversation)
