--- src/kenai/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class KenaiConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "kenai"
