--- src/kenai/serializers.py
@@ -1,5 +1,0 @@
-from rest_framework import serializers
-
-
-class KenaiMessageSerializer(serializers.Serializer):
-    message = serializers.CharField()
