--- src/kenai/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import OpenaiAPIView
-
-urlpatterns = [
-    path("chat/", OpenaiAPIView.as_view(), name="kenai-openai-chat"),
-]
