+++ src/requirements.txt
@@ -18,4 +18,3 @@ django-modeltranslation==0.18.11
 celery==5.4.0
 channels_redis==4.2.0
 daphne==4.0.0
-openai==1.35.10
